List exported functions and their arguments

Generate a list of the functions exported by a package, along with their arguments, and the default values of those arguments.

Usage

pkg_fn_info(pkg)

Arguments

pkg

A character naming an installed package.

Value

A named list of named pairlists. The names of the outer list are the functions, and the names of each contained pairlist are the arguments. The values of the pairlists are the default values of the arguments.

Details

Note: While you can get values for "base" as if it is a package, primitive functions will show NULL as their arguments.

Examples

base_fns <- pkg_fn_info("base")
# Note that we do not know information about the arguments of primatives.
base_fns$is.function
#> NULL

pkg_fn_info("utils")
#> $aspell_package_Rd_files
#> $aspell_package_Rd_files$dir
#> 
#> 
#> $aspell_package_Rd_files$drop
#> c("\\author", "\\references")
#> 
#> $aspell_package_Rd_files$control
#> list()
#> 
#> $aspell_package_Rd_files$program
#> NULL
#> 
#> $aspell_package_Rd_files$dictionaries
#> character()
#> 
#> 
#> $vi
#> $vi$name
#> NULL
#> 
#> $vi$file
#> [1] ""
#> 
#> 
#> $read.table
#> $read.table$file
#> 
#> 
#> $read.table$header
#> [1] FALSE
#> 
#> $read.table$sep
#> [1] ""
#> 
#> $read.table$quote
#> [1] "\"'"
#> 
#> $read.table$dec
#> [1] "."
#> 
#> $read.table$numerals
#> c("allow.loss", "warn.loss", "no.loss")
#> 
#> $read.table$row.names
#> 
#> 
#> $read.table$col.names
#> 
#> 
#> $read.table$as.is
#> !stringsAsFactors
#> 
#> $read.table$na.strings
#> [1] "NA"
#> 
#> $read.table$colClasses
#> [1] NA
#> 
#> $read.table$nrows
#> -1
#> 
#> $read.table$skip
#> [1] 0
#> 
#> $read.table$check.names
#> [1] TRUE
#> 
#> $read.table$fill
#> !blank.lines.skip
#> 
#> $read.table$strip.white
#> [1] FALSE
#> 
#> $read.table$blank.lines.skip
#> [1] TRUE
#> 
#> $read.table$comment.char
#> [1] "#"
#> 
#> $read.table$allowEscapes
#> [1] FALSE
#> 
#> $read.table$flush
#> [1] FALSE
#> 
#> $read.table$stringsAsFactors
#> [1] FALSE
#> 
#> $read.table$fileEncoding
#> [1] ""
#> 
#> $read.table$encoding
#> [1] "unknown"
#> 
#> $read.table$text
#> 
#> 
#> $read.table$skipNul
#> [1] FALSE
#> 
#> 
#> $URLdecode
#> $URLdecode$URL
#> 
#> 
#> 
#> $rc.status
#> NULL
#> 
#> $write.csv
#> $write.csv$...
#> 
#> 
#> 
#> $RweaveLatexOptions
#> $RweaveLatexOptions$options
#> 
#> 
#> 
#> $formatUL
#> $formatUL$x
#> 
#> 
#> $formatUL$label
#> [1] "*"
#> 
#> $formatUL$offset
#> [1] 0
#> 
#> $formatUL$width
#> 0.9 * getOption("width")
#> 
#> 
#> $prompt
#> $prompt$object
#> 
#> 
#> $prompt$filename
#> NULL
#> 
#> $prompt$name
#> NULL
#> 
#> $prompt$...
#> 
#> 
#> 
#> $upgrade
#> $upgrade$object
#> 
#> 
#> $upgrade$...
#> 
#> 
#> 
#> $RShowDoc
#> $RShowDoc$what
#> 
#> 
#> $RShowDoc$type
#> c("pdf", "html", "txt")
#> 
#> $RShowDoc$package
#> 
#> 
#> 
#> $argsAnywhere
#> $argsAnywhere$x
#> 
#> 
#> 
#> $read.delim
#> $read.delim$file
#> 
#> 
#> $read.delim$header
#> [1] TRUE
#> 
#> $read.delim$sep
#> [1] "\t"
#> 
#> $read.delim$quote
#> [1] "\""
#> 
#> $read.delim$dec
#> [1] "."
#> 
#> $read.delim$fill
#> [1] TRUE
#> 
#> $read.delim$comment.char
#> [1] ""
#> 
#> $read.delim$...
#> 
#> 
#> 
#> $de.ncols
#> $de.ncols$inlist
#> 
#> 
#> 
#> $hashtab
#> $hashtab$type
#> c("identical", "address")
#> 
#> $hashtab$size
#> 
#> 
#> 
#> $SweaveSyntConv
#> $SweaveSyntConv$file
#> 
#> 
#> $SweaveSyntConv$syntax
#> 
#> 
#> $SweaveSyntConv$output
#> NULL
#> 
#> 
#> $new.packages
#> $new.packages$lib.loc
#> NULL
#> 
#> $new.packages$repos
#> getOption("repos")
#> 
#> $new.packages$contriburl
#> contrib.url(repos, type)
#> 
#> $new.packages$instPkgs
#> installed.packages(lib.loc = lib.loc, ...)
#> 
#> $new.packages$method
#> 
#> 
#> $new.packages$available
#> NULL
#> 
#> $new.packages$ask
#> [1] FALSE
#> 
#> $new.packages$...
#> 
#> 
#> $new.packages$type
#> getOption("pkgType")
#> 
#> 
#> $read.socket
#> $read.socket$socket
#> 
#> 
#> $read.socket$maxlen
#> [1] 256
#> 
#> $read.socket$loop
#> [1] FALSE
#> 
#> 
#> $personList
#> $personList$...
#> 
#> 
#> 
#> $write.table
#> $write.table$x
#> 
#> 
#> $write.table$file
#> [1] ""
#> 
#> $write.table$append
#> [1] FALSE
#> 
#> $write.table$quote
#> [1] TRUE
#> 
#> $write.table$sep
#> [1] " "
#> 
#> $write.table$eol
#> [1] "\n"
#> 
#> $write.table$na
#> [1] "NA"
#> 
#> $write.table$dec
#> [1] "."
#> 
#> $write.table$row.names
#> [1] TRUE
#> 
#> $write.table$col.names
#> [1] TRUE
#> 
#> $write.table$qmethod
#> c("escape", "double")
#> 
#> $write.table$fileEncoding
#> [1] ""
#> 
#> 
#> $read.delim2
#> $read.delim2$file
#> 
#> 
#> $read.delim2$header
#> [1] TRUE
#> 
#> $read.delim2$sep
#> [1] "\t"
#> 
#> $read.delim2$quote
#> [1] "\""
#> 
#> $read.delim2$dec
#> [1] ","
#> 
#> $read.delim2$fill
#> [1] TRUE
#> 
#> $read.delim2$comment.char
#> [1] ""
#> 
#> $read.delim2$...
#> 
#> 
#> 
#> $aspell_package_vignettes
#> $aspell_package_vignettes$dir
#> 
#> 
#> $aspell_package_vignettes$control
#> list()
#> 
#> $aspell_package_vignettes$program
#> NULL
#> 
#> $aspell_package_vignettes$dictionaries
#> character()
#> 
#> 
#> $SweaveHooks
#> $SweaveHooks$options
#> 
#> 
#> $SweaveHooks$run
#> [1] FALSE
#> 
#> $SweaveHooks$envir
#> .GlobalEnv
#> 
#> 
#> $hasName
#> $hasName$x
#> 
#> 
#> $hasName$name
#> 
#> 
#> 
#> $assignInNamespace
#> $assignInNamespace$x
#> 
#> 
#> $assignInNamespace$value
#> 
#> 
#> $assignInNamespace$ns
#> 
#> 
#> $assignInNamespace$pos
#> -1
#> 
#> $assignInNamespace$envir
#> as.environment(pos)
#> 
#> 
#> $is.relistable
#> $is.relistable$x
#> 
#> 
#> 
#> $.DollarNames
#> $.DollarNames$x
#> 
#> 
#> $.DollarNames$pattern
#> 
#> 
#> 
#> $toBibtex
#> $toBibtex$object
#> 
#> 
#> $toBibtex$...
#> 
#> 
#> 
#> $alarm
#> NULL
#> 
#> $RweaveLatexSetup
#> $RweaveLatexSetup$file
#> 
#> 
#> $RweaveLatexSetup$syntax
#> 
#> 
#> $RweaveLatexSetup$output
#> NULL
#> 
#> $RweaveLatexSetup$quiet
#> [1] FALSE
#> 
#> $RweaveLatexSetup$debug
#> [1] FALSE
#> 
#> $RweaveLatexSetup$stylepath
#> 
#> 
#> $RweaveLatexSetup$...
#> 
#> 
#> 
#> $checkCRAN
#> $checkCRAN$method
#> 
#> 
#> 
#> $setRepositories
#> $setRepositories$graphics
#> getOption("menu.graphics")
#> 
#> $setRepositories$ind
#> NULL
#> 
#> $setRepositories$addURLs
#> character()
#> 
#> 
#> $warnErrList
#> $warnErrList$x
#> 
#> 
#> $warnErrList$warn
#> [1] TRUE
#> 
#> $warnErrList$errValue
#> NULL
#> 
#> 
#> $sethash
#> $sethash$h
#> 
#> 
#> $sethash$key
#> 
#> 
#> $sethash$value
#> 
#> 
#> 
#> $dump.frames
#> $dump.frames$dumpto
#> [1] "last.dump"
#> 
#> $dump.frames$to.file
#> [1] FALSE
#> 
#> $dump.frames$include.GlobalEnv
#> [1] FALSE
#> 
#> 
#> $Rprof
#> $Rprof$filename
#> [1] "Rprof.out"
#> 
#> $Rprof$append
#> [1] FALSE
#> 
#> $Rprof$interval
#> [1] 0.02
#> 
#> $Rprof$memory.profiling
#> [1] FALSE
#> 
#> $Rprof$gc.profiling
#> [1] FALSE
#> 
#> $Rprof$line.profiling
#> [1] FALSE
#> 
#> $Rprof$filter.callframes
#> [1] FALSE
#> 
#> $Rprof$numfiles
#> [1] 100
#> 
#> $Rprof$bufsize
#> [1] 10000
#> 
#> 
#> $clrhash
#> $clrhash$h
#> 
#> 
#> 
#> $sessionInfo
#> $sessionInfo$package
#> NULL
#> 
#> 
#> $count.fields
#> $count.fields$file
#> 
#> 
#> $count.fields$sep
#> [1] ""
#> 
#> $count.fields$quote
#> [1] "\"'"
#> 
#> $count.fields$skip
#> [1] 0
#> 
#> $count.fields$blank.lines.skip
#> [1] TRUE
#> 
#> $count.fields$comment.char
#> [1] "#"
#> 
#> 
#> $create.post
#> $create.post$instructions
#> character()
#> 
#> $create.post$description
#> [1] "post"
#> 
#> $create.post$subject
#> [1] ""
#> 
#> $create.post$method
#> getOption("mailer")
#> 
#> $create.post$address
#> [1] "the relevant mailing list"
#> 
#> $create.post$ccaddress
#> getOption("ccaddress", "")
#> 
#> $create.post$filename
#> [1] "R.post"
#> 
#> $create.post$info
#> character()
#> 
#> 
#> $tail.matrix
#> $tail.matrix$x
#> 
#> 
#> $tail.matrix$n
#> [1] 6
#> 
#> $tail.matrix$keepnums
#> [1] TRUE
#> 
#> $tail.matrix$addrownums
#> 
#> 
#> $tail.matrix$...
#> 
#> 
#> 
#> $person
#> $person$given
#> NULL
#> 
#> $person$family
#> NULL
#> 
#> $person$middle
#> NULL
#> 
#> $person$email
#> NULL
#> 
#> $person$role
#> NULL
#> 
#> $person$comment
#> NULL
#> 
#> $person$first
#> NULL
#> 
#> $person$last
#> NULL
#> 
#> 
#> $promptPackage
#> $promptPackage$package
#> 
#> 
#> $promptPackage$lib.loc
#> NULL
#> 
#> $promptPackage$filename
#> NULL
#> 
#> $promptPackage$name
#> NULL
#> 
#> $promptPackage$final
#> [1] FALSE
#> 
#> 
#> $install.packages
#> $install.packages$pkgs
#> 
#> 
#> $install.packages$lib
#> 
#> 
#> $install.packages$repos
#> getOption("repos")
#> 
#> $install.packages$contriburl
#> contrib.url(repos, type)
#> 
#> $install.packages$method
#> 
#> 
#> $install.packages$available
#> NULL
#> 
#> $install.packages$destdir
#> NULL
#> 
#> $install.packages$dependencies
#> [1] NA
#> 
#> $install.packages$type
#> getOption("pkgType")
#> 
#> $install.packages$configure.args
#> getOption("configure.args")
#> 
#> $install.packages$configure.vars
#> getOption("configure.vars")
#> 
#> $install.packages$clean
#> [1] FALSE
#> 
#> $install.packages$Ncpus
#> getOption("Ncpus", 1L)
#> 
#> $install.packages$verbose
#> getOption("verbose")
#> 
#> $install.packages$libs_only
#> [1] FALSE
#> 
#> $install.packages$INSTALL_opts
#> 
#> 
#> $install.packages$quiet
#> [1] FALSE
#> 
#> $install.packages$keep_outputs
#> [1] FALSE
#> 
#> $install.packages$...
#> 
#> 
#> 
#> $news
#> $news$query
#> 
#> 
#> $news$package
#> [1] "R"
#> 
#> $news$lib.loc
#> NULL
#> 
#> $news$format
#> NULL
#> 
#> $news$reader
#> NULL
#> 
#> $news$db
#> NULL
#> 
#> 
#> $URLencode
#> $URLencode$URL
#> 
#> 
#> $URLencode$reserved
#> [1] FALSE
#> 
#> $URLencode$repeated
#> [1] FALSE
#> 
#> 
#> $type.convert
#> $type.convert$x
#> 
#> 
#> $type.convert$...
#> 
#> 
#> 
#> $help.start
#> $help.start$update
#> [1] FALSE
#> 
#> $help.start$gui
#> [1] "irrelevant"
#> 
#> $help.start$browser
#> getOption("browser")
#> 
#> $help.start$remote
#> NULL
#> 
#> 
#> $de.setup
#> $de.setup$ilist
#> 
#> 
#> $de.setup$list.names
#> 
#> 
#> $de.setup$incols
#> 
#> 
#> 
#> $aregexec
#> $aregexec$pattern
#> 
#> 
#> $aregexec$text
#> 
#> 
#> $aregexec$max.distance
#> [1] 0.1
#> 
#> $aregexec$costs
#> NULL
#> 
#> $aregexec$ignore.case
#> [1] FALSE
#> 
#> $aregexec$fixed
#> [1] FALSE
#> 
#> $aregexec$useBytes
#> [1] FALSE
#> 
#> 
#> $rc.getOption
#> $rc.getOption$name
#> 
#> 
#> 
#> $maphash
#> $maphash$h
#> 
#> 
#> $maphash$FUN
#> 
#> 
#> 
#> $de.restore
#> $de.restore$inlist
#> 
#> 
#> $de.restore$ncols
#> 
#> 
#> $de.restore$coltypes
#> 
#> 
#> $de.restore$argnames
#> 
#> 
#> $de.restore$args
#> 
#> 
#> 
#> $file.edit
#> $file.edit$...
#> 
#> 
#> $file.edit$title
#> file
#> 
#> $file.edit$editor
#> getOption("editor")
#> 
#> $file.edit$fileEncoding
#> [1] ""
#> 
#> 
#> $remhash
#> $remhash$h
#> 
#> 
#> $remhash$key
#> 
#> 
#> 
#> $citHeader
#> $citHeader$...
#> 
#> 
#> 
#> $fix
#> $fix$x
#> 
#> 
#> $fix$...
#> 
#> 
#> 
#> $makeRweaveLatexCodeRunner
#> $makeRweaveLatexCodeRunner$evalFunc
#> RweaveEvalWithOpt
#> 
#> 
#> $.RtangleCodeLabel
#> $.RtangleCodeLabel$chunk
#> 
#> 
#> 
#> $write.socket
#> $write.socket$socket
#> 
#> 
#> $write.socket$string
#> 
#> 
#> 
#> $unzip
#> $unzip$zipfile
#> 
#> 
#> $unzip$files
#> NULL
#> 
#> $unzip$list
#> [1] FALSE
#> 
#> $unzip$overwrite
#> [1] TRUE
#> 
#> $unzip$junkpaths
#> [1] FALSE
#> 
#> $unzip$exdir
#> [1] "."
#> 
#> $unzip$unzip
#> [1] "internal"
#> 
#> $unzip$setTimes
#> [1] FALSE
#> 
#> 
#> $Sweave
#> $Sweave$file
#> 
#> 
#> $Sweave$driver
#> RweaveLatex()
#> 
#> $Sweave$syntax
#> getOption("SweaveSyntax")
#> 
#> $Sweave$encoding
#> [1] ""
#> 
#> $Sweave$...
#> 
#> 
#> 
#> $nsl
#> $nsl$hostname
#> 
#> 
#> 
#> $read.fwf
#> $read.fwf$file
#> 
#> 
#> $read.fwf$widths
#> 
#> 
#> $read.fwf$header
#> [1] FALSE
#> 
#> $read.fwf$sep
#> [1] "\t"
#> 
#> $read.fwf$skip
#> [1] 0
#> 
#> $read.fwf$row.names
#> 
#> 
#> $read.fwf$col.names
#> 
#> 
#> $read.fwf$n
#> -1L
#> 
#> $read.fwf$buffersize
#> [1] 2000
#> 
#> $read.fwf$fileEncoding
#> [1] ""
#> 
#> $read.fwf$...
#> 
#> 
#> 
#> $RweaveEvalWithOpt
#> $RweaveEvalWithOpt$expr
#> 
#> 
#> $RweaveEvalWithOpt$options
#> 
#> 
#> 
#> $limitedLabels
#> $limitedLabels$value
#> 
#> 
#> $limitedLabels$maxwidth
#> getOption("width") - 5L
#> 
#> 
#> $readCitationFile
#> $readCitationFile$file
#> 
#> 
#> $readCitationFile$meta
#> NULL
#> 
#> 
#> $`?`
#> $`?`$e1
#> 
#> 
#> $`?`$e2
#> 
#> 
#> 
#> $process.events
#> NULL
#> 
#> $as.relistable
#> $as.relistable$x
#> 
#> 
#> 
#> $getParseText
#> $getParseText$parseData
#> 
#> 
#> $getParseText$id
#> 
#> 
#> 
#> $isS3stdGeneric
#> $isS3stdGeneric$f
#> 
#> 
#> 
#> $modifyList
#> $modifyList$x
#> 
#> 
#> $modifyList$val
#> 
#> 
#> $modifyList$keep.null
#> [1] FALSE
#> 
#> 
#> $maintainer
#> $maintainer$pkg
#> 
#> 
#> 
#> $available.packages
#> $available.packages$contriburl
#> contrib.url(repos, type)
#> 
#> $available.packages$method
#> 
#> 
#> $available.packages$fields
#> NULL
#> 
#> $available.packages$type
#> getOption("pkgType")
#> 
#> $available.packages$filters
#> NULL
#> 
#> $available.packages$repos
#> getOption("repos")
#> 
#> $available.packages$ignore_repo_cache
#> [1] FALSE
#> 
#> $available.packages$max_repo_cache_age
#> 
#> 
#> $available.packages$quiet
#> [1] TRUE
#> 
#> $available.packages$...
#> 
#> 
#> 
#> $citeNatbib
#> $citeNatbib$keys
#> 
#> 
#> $citeNatbib$bib
#> 
#> 
#> $citeNatbib$textual
#> [1] FALSE
#> 
#> $citeNatbib$before
#> NULL
#> 
#> $citeNatbib$after
#> NULL
#> 
#> $citeNatbib$mode
#> c("authoryear", "numbers", "super")
#> 
#> $citeNatbib$abbreviate
#> [1] TRUE
#> 
#> $citeNatbib$longnamesfirst
#> [1] TRUE
#> 
#> $citeNatbib$bibpunct
#> c("(", ")", ";", "a", "", ",")
#> 
#> $citeNatbib$previous
#> 
#> 
#> 
#> $rc.options
#> $rc.options$...
#> 
#> 
#> 
#> $Rtangle
#> NULL
#> 
#> $setBreakpoint
#> $setBreakpoint$srcfile
#> 
#> 
#> $setBreakpoint$line
#> 
#> 
#> $setBreakpoint$nameonly
#> [1] TRUE
#> 
#> $setBreakpoint$envir
#> parent.frame()
#> 
#> $setBreakpoint$lastenv
#> 
#> 
#> $setBreakpoint$verbose
#> [1] TRUE
#> 
#> $setBreakpoint$tracer
#> 
#> 
#> $setBreakpoint$print
#> [1] FALSE
#> 
#> $setBreakpoint$clear
#> [1] FALSE
#> 
#> $setBreakpoint$...
#> 
#> 
#> 
#> $head.matrix
#> $head.matrix$x
#> 
#> 
#> $head.matrix$n
#> [1] 6
#> 
#> $head.matrix$...
#> 
#> 
#> 
#> $getTxtProgressBar
#> $getTxtProgressBar$pb
#> 
#> 
#> 
#> $tail
#> $tail$x
#> 
#> 
#> $tail$...
#> 
#> 
#> 
#> $changedFiles
#> $changedFiles$before
#> 
#> 
#> $changedFiles$after
#> 
#> 
#> $changedFiles$path
#> before$path
#> 
#> $changedFiles$timestamp
#> before$timestamp
#> 
#> $changedFiles$check.file.info
#> c("size", "isdir", "mode", "mtime")
#> 
#> $changedFiles$md5sum
#> before$md5sum
#> 
#> $changedFiles$digest
#> before$digest
#> 
#> $changedFiles$full.names
#> before$full.names
#> 
#> $changedFiles$...
#> 
#> 
#> 
#> $RweaveLatexFinish
#> $RweaveLatexFinish$object
#> 
#> 
#> $RweaveLatexFinish$error
#> [1] FALSE
#> 
#> 
#> $numhash
#> $numhash$h
#> 
#> 
#> 
#> $getAnywhere
#> $getAnywhere$x
#> 
#> 
#> 
#> $browseEnv
#> $browseEnv$envir
#> .GlobalEnv
#> 
#> $browseEnv$pattern
#> 
#> 
#> $browseEnv$excludepatt
#> [1] "^last\\.warning"
#> 
#> $browseEnv$html
#> .Platform$GUI != "AQUA"
#> 
#> $browseEnv$expanded
#> [1] TRUE
#> 
#> $browseEnv$properties
#> NULL
#> 
#> $browseEnv$main
#> NULL
#> 
#> $browseEnv$debugMe
#> [1] FALSE
#> 
#> 
#> $object.size
#> $object.size$x
#> 
#> 
#> 
#> $globalVariables
#> $globalVariables$names
#> 
#> 
#> $globalVariables$package
#> 
#> 
#> $globalVariables$add
#> [1] TRUE
#> 
#> 
#> $packageVersion
#> $packageVersion$pkg
#> 
#> 
#> $packageVersion$lib.loc
#> NULL
#> 
#> 
#> $help.search
#> $help.search$pattern
#> 
#> 
#> $help.search$fields
#> c("alias", "concept", "title")
#> 
#> $help.search$apropos
#> 
#> 
#> $help.search$keyword
#> 
#> 
#> $help.search$whatis
#> 
#> 
#> $help.search$ignore.case
#> [1] TRUE
#> 
#> $help.search$package
#> NULL
#> 
#> $help.search$lib.loc
#> NULL
#> 
#> $help.search$help.db
#> getOption("help.db")
#> 
#> $help.search$verbose
#> getOption("verbose")
#> 
#> $help.search$rebuild
#> [1] FALSE
#> 
#> $help.search$agrep
#> NULL
#> 
#> $help.search$use_UTF8
#> [1] FALSE
#> 
#> $help.search$types
#> getOption("help.search.types")
#> 
#> 
#> $package.skeleton
#> $package.skeleton$name
#> [1] "anRpackage"
#> 
#> $package.skeleton$list
#> character()
#> 
#> $package.skeleton$environment
#> .GlobalEnv
#> 
#> $package.skeleton$path
#> [1] "."
#> 
#> $package.skeleton$force
#> [1] FALSE
#> 
#> $package.skeleton$code_files
#> character()
#> 
#> $package.skeleton$encoding
#> [1] "unknown"
#> 
#> 
#> $summaryRprof
#> $summaryRprof$filename
#> [1] "Rprof.out"
#> 
#> $summaryRprof$chunksize
#> [1] 5000
#> 
#> $summaryRprof$memory
#> c("none", "both", "tseries", "stats")
#> 
#> $summaryRprof$lines
#> c("hide", "show", "both")
#> 
#> $summaryRprof$index
#> [1] 2
#> 
#> $summaryRprof$diff
#> [1] TRUE
#> 
#> $summaryRprof$exclude
#> NULL
#> 
#> $summaryRprof$basenames
#> [1] 1
#> 
#> 
#> $setTxtProgressBar
#> $setTxtProgressBar$pb
#> 
#> 
#> $setTxtProgressBar$value
#> 
#> 
#> $setTxtProgressBar$title
#> NULL
#> 
#> $setTxtProgressBar$label
#> NULL
#> 
#> 
#> $timestamp
#> $timestamp$stamp
#> date()
#> 
#> $timestamp$prefix
#> [1] "##------ "
#> 
#> $timestamp$suffix
#> [1] " ------##"
#> 
#> $timestamp$quiet
#> [1] FALSE
#> 
#> 
#> $getSrcref
#> $getSrcref$x
#> 
#> 
#> 
#> $stack
#> $stack$x
#> 
#> 
#> $stack$...
#> 
#> 
#> 
#> $is.hashtab
#> $is.hashtab$x
#> 
#> 
#> 
#> $capture.output
#> $capture.output$...
#> 
#> 
#> $capture.output$file
#> NULL
#> 
#> $capture.output$append
#> [1] FALSE
#> 
#> $capture.output$type
#> c("output", "message")
#> 
#> $capture.output$split
#> [1] FALSE
#> 
#> 
#> $bug.report
#> $bug.report$subject
#> [1] ""
#> 
#> $bug.report$address
#> 
#> 
#> $bug.report$file
#> [1] "R.bug.report"
#> 
#> $bug.report$package
#> NULL
#> 
#> $bug.report$lib.loc
#> NULL
#> 
#> $bug.report$...
#> 
#> 
#> 
#> $rc.settings
#> $rc.settings$ops
#> 
#> 
#> $rc.settings$ns
#> 
#> 
#> $rc.settings$args
#> 
#> 
#> $rc.settings$dots
#> 
#> 
#> $rc.settings$func
#> 
#> 
#> $rc.settings$ipck
#> 
#> 
#> $rc.settings$S3
#> 
#> 
#> $rc.settings$data
#> 
#> 
#> $rc.settings$help
#> 
#> 
#> $rc.settings$argdb
#> 
#> 
#> $rc.settings$fuzzy
#> 
#> 
#> $rc.settings$quotes
#> 
#> 
#> $rc.settings$files
#> 
#> 
#> 
#> $fileSnapshot
#> $fileSnapshot$path
#> [1] "."
#> 
#> $fileSnapshot$file.info
#> [1] TRUE
#> 
#> $fileSnapshot$timestamp
#> NULL
#> 
#> $fileSnapshot$md5sum
#> [1] FALSE
#> 
#> $fileSnapshot$digest
#> NULL
#> 
#> $fileSnapshot$full.names
#> length(path) > 1
#> 
#> $fileSnapshot$...
#> 
#> 
#> 
#> $hsearch_db_keywords
#> $hsearch_db_keywords$db
#> hsearch_db()
#> 
#> 
#> $rtags
#> $rtags$path
#> [1] "."
#> 
#> $rtags$pattern
#> [1] "\\.[RrSs]$"
#> 
#> $rtags$recursive
#> [1] FALSE
#> 
#> $rtags$src
#> list.files(path = path, pattern = pattern, full.names = TRUE, 
#>     recursive = recursive)
#> 
#> $rtags$keep.re
#> NULL
#> 
#> $rtags$ofile
#> [1] ""
#> 
#> $rtags$append
#> [1] FALSE
#> 
#> $rtags$verbose
#> getOption("verbose")
#> 
#> $rtags$type
#> c("etags", "ctags")
#> 
#> 
#> $browseVignettes
#> $browseVignettes$package
#> NULL
#> 
#> $browseVignettes$lib.loc
#> NULL
#> 
#> $browseVignettes$all
#> [1] TRUE
#> 
#> 
#> $contrib.url
#> $contrib.url$repos
#> 
#> 
#> $contrib.url$type
#> getOption("pkgType")
#> 
#> 
#> $strOptions
#> $strOptions$strict.width
#> [1] "no"
#> 
#> $strOptions$digits.d
#> [1] 3
#> 
#> $strOptions$vec.len
#> [1] 4
#> 
#> $strOptions$list.len
#> [1] 99
#> 
#> $strOptions$deparse.lines
#> NULL
#> 
#> $strOptions$drop.deparse.attr
#> [1] TRUE
#> 
#> $strOptions$formatNum
#> function(x, ...) format(x, trim = TRUE, drop0trailing = TRUE, 
#>     ...)
#> 
#> 
#> $RweaveLatex
#> NULL
#> 
#> $packageDescription
#> $packageDescription$pkg
#> 
#> 
#> $packageDescription$lib.loc
#> NULL
#> 
#> $packageDescription$fields
#> NULL
#> 
#> $packageDescription$drop
#> [1] TRUE
#> 
#> $packageDescription$encoding
#> [1] ""
#> 
#> 
#> $flush.console
#> NULL
#> 
#> $Rprofmem
#> $Rprofmem$filename
#> [1] "Rprofmem.out"
#> 
#> $Rprofmem$append
#> [1] FALSE
#> 
#> $Rprofmem$threshold
#> [1] 0
#> 
#> 
#> $asDateBuilt
#> $asDateBuilt$built
#> 
#> 
#> 
#> $compareVersion
#> $compareVersion$a
#> 
#> 
#> $compareVersion$b
#> 
#> 
#> 
#> $charClass
#> $charClass$x
#> 
#> 
#> $charClass$class
#> 
#> 
#> 
#> $as.personList
#> $as.personList$x
#> 
#> 
#> 
#> $getFromNamespace
#> $getFromNamespace$x
#> 
#> 
#> $getFromNamespace$ns
#> 
#> 
#> $getFromNamespace$pos
#> -1
#> 
#> $getFromNamespace$envir
#> as.environment(pos)
#> 
#> 
#> $recover
#> NULL
#> 
#> $packageStatus
#> $packageStatus$lib.loc
#> NULL
#> 
#> $packageStatus$repositories
#> NULL
#> 
#> $packageStatus$method
#> 
#> 
#> $packageStatus$type
#> getOption("pkgType")
#> 
#> $packageStatus$...
#> 
#> 
#> 
#> $read.csv2
#> $read.csv2$file
#> 
#> 
#> $read.csv2$header
#> [1] TRUE
#> 
#> $read.csv2$sep
#> [1] ";"
#> 
#> $read.csv2$quote
#> [1] "\""
#> 
#> $read.csv2$dec
#> [1] ","
#> 
#> $read.csv2$fill
#> [1] TRUE
#> 
#> $read.csv2$comment.char
#> [1] ""
#> 
#> $read.csv2$...
#> 
#> 
#> 
#> $demo
#> $demo$topic
#> 
#> 
#> $demo$package
#> NULL
#> 
#> $demo$lib.loc
#> NULL
#> 
#> $demo$character.only
#> [1] FALSE
#> 
#> $demo$verbose
#> getOption("verbose")
#> 
#> $demo$type
#> c("console", "html")
#> 
#> $demo$echo
#> [1] TRUE
#> 
#> $demo$ask
#> getOption("demo.ask")
#> 
#> $demo$encoding
#> getOption("encoding")
#> 
#> 
#> $edit
#> $edit$name
#> 
#> 
#> $edit$...
#> 
#> 
#> 
#> $aspell_package_R_files
#> $aspell_package_R_files$dir
#> 
#> 
#> $aspell_package_R_files$ignore
#> character()
#> 
#> $aspell_package_R_files$control
#> list()
#> 
#> $aspell_package_R_files$program
#> NULL
#> 
#> $aspell_package_R_files$dictionaries
#> character()
#> 
#> 
#> $isS3method
#> $isS3method$method
#> 
#> 
#> $isS3method$f
#> 
#> 
#> $isS3method$class
#> 
#> 
#> $isS3method$envir
#> parent.frame()
#> 
#> 
#> $de
#> $de$...
#> 
#> 
#> $de$Modes
#> list()
#> 
#> $de$Names
#> NULL
#> 
#> 
#> $packageDate
#> $packageDate$pkg
#> 
#> 
#> $packageDate$lib.loc
#> NULL
#> 
#> $packageDate$date.fields
#> c("Date", "Packaged", "Date/Publication", "Built")
#> 
#> $packageDate$tryFormats
#> c("%Y-%m-%d", "%Y/%m/%d", "%D", "%m/%d/%y")
#> 
#> $packageDate$desc
#> packageDescription(pkg, lib.loc = lib.loc, fields = date.fields)
#> 
#> 
#> $RSiteSearch
#> $RSiteSearch$string
#> 
#> 
#> $RSiteSearch$restrict
#> c("functions", "descriptions", "news", "Rfunctions", "Rmanuals", 
#>     "READMEs", "views", "vignettes")
#> 
#> $RSiteSearch$format
#> 
#> 
#> $RSiteSearch$sortby
#> c("score", "date:late", "date:early", "subject", "subject:descending", 
#>     "size", "size:descending")
#> 
#> $RSiteSearch$matchesPerPage
#> [1] 20
#> 
#> $RSiteSearch$words
#> c("all", "any")
#> 
#> 
#> $read.DIF
#> $read.DIF$file
#> 
#> 
#> $read.DIF$header
#> [1] FALSE
#> 
#> $read.DIF$dec
#> [1] "."
#> 
#> $read.DIF$numerals
#> c("allow.loss", "warn.loss", "no.loss")
#> 
#> $read.DIF$row.names
#> 
#> 
#> $read.DIF$col.names
#> 
#> 
#> $read.DIF$as.is
#> !stringsAsFactors
#> 
#> $read.DIF$na.strings
#> [1] "NA"
#> 
#> $read.DIF$colClasses
#> [1] NA
#> 
#> $read.DIF$nrows
#> -1
#> 
#> $read.DIF$skip
#> [1] 0
#> 
#> $read.DIF$check.names
#> [1] TRUE
#> 
#> $read.DIF$blank.lines.skip
#> [1] TRUE
#> 
#> $read.DIF$stringsAsFactors
#> [1] FALSE
#> 
#> $read.DIF$transpose
#> [1] FALSE
#> 
#> $read.DIF$fileEncoding
#> [1] ""
#> 
#> 
#> $hsearch_db_concepts
#> $hsearch_db_concepts$db
#> hsearch_db()
#> 
#> 
#> $getSrcLocation
#> $getSrcLocation$x
#> 
#> 
#> $getSrcLocation$which
#> c("line", "column", "byte", "parse")
#> 
#> $getSrcLocation$first
#> [1] TRUE
#> 
#> 
#> $lsf.str
#> $lsf.str$pos
#> -1
#> 
#> $lsf.str$envir
#> 
#> 
#> $lsf.str$...
#> 
#> 
#> 
#> $tar
#> $tar$tarfile
#> 
#> 
#> $tar$files
#> NULL
#> 
#> $tar$compression
#> c("none", "gzip", "bzip2", "xz")
#> 
#> $tar$compression_level
#> [1] 6
#> 
#> $tar$tar
#> Sys.getenv("tar")
#> 
#> $tar$extra_flags
#> [1] ""
#> 
#> 
#> $help.request
#> $help.request$subject
#> [1] ""
#> 
#> $help.request$address
#> [1] "r-help@R-project.org"
#> 
#> $help.request$file
#> [1] "R.help.request"
#> 
#> $help.request$...
#> 
#> 
#> 
#> $cite
#> $cite$keys
#> 
#> 
#> $cite$bib
#> 
#> 
#> $cite$...
#> 
#> 
#> 
#> $.S3methods
#> $.S3methods$generic.function
#> 
#> 
#> $.S3methods$class
#> 
#> 
#> $.S3methods$envir
#> parent.frame()
#> 
#> 
#> $select.list
#> $select.list$choices
#> 
#> 
#> $select.list$preselect
#> NULL
#> 
#> $select.list$multiple
#> [1] FALSE
#> 
#> $select.list$title
#> NULL
#> 
#> $select.list$graphics
#> getOption("menu.graphics")
#> 
#> 
#> $getCRANmirrors
#> $getCRANmirrors$all
#> [1] FALSE
#> 
#> $getCRANmirrors$local.only
#> [1] FALSE
#> 
#> 
#> $relist
#> $relist$flesh
#> 
#> 
#> $relist$skeleton
#> attr(flesh, "skeleton")
#> 
#> 
#> $ls.str
#> $ls.str$pos
#> -1
#> 
#> $ls.str$name
#> 
#> 
#> $ls.str$envir
#> 
#> 
#> $ls.str$all.names
#> [1] FALSE
#> 
#> $ls.str$pattern
#> 
#> 
#> $ls.str$mode
#> [1] "any"
#> 
#> 
#> $localeToCharset
#> $localeToCharset$locale
#> Sys.getlocale("LC_CTYPE")
#> 
#> 
#> $download.packages
#> $download.packages$pkgs
#> 
#> 
#> $download.packages$destdir
#> 
#> 
#> $download.packages$available
#> NULL
#> 
#> $download.packages$repos
#> getOption("repos")
#> 
#> $download.packages$contriburl
#> contrib.url(repos, type)
#> 
#> $download.packages$method
#> 
#> 
#> $download.packages$type
#> getOption("pkgType")
#> 
#> $download.packages$...
#> 
#> 
#> 
#> $debugger
#> $debugger$dump
#> last.dump
#> 
#> 
#> $chooseCRANmirror
#> $chooseCRANmirror$graphics
#> getOption("menu.graphics")
#> 
#> $chooseCRANmirror$ind
#> NULL
#> 
#> $chooseCRANmirror$local.only
#> [1] FALSE
#> 
#> 
#> $savehistory
#> $savehistory$file
#> [1] ".Rhistory"
#> 
#> 
#> $citEntry
#> $citEntry$entry
#> 
#> 
#> $citEntry$textVersion
#> 
#> 
#> $citEntry$header
#> NULL
#> 
#> $citEntry$footer
#> NULL
#> 
#> $citEntry$...
#> 
#> 
#> 
#> $RweaveTryStop
#> $RweaveTryStop$err
#> 
#> 
#> $RweaveTryStop$options
#> 
#> 
#> 
#> $close.socket
#> $close.socket$socket
#> 
#> 
#> $close.socket$...
#> 
#> 
#> 
#> $vignette
#> $vignette$topic
#> 
#> 
#> $vignette$package
#> NULL
#> 
#> $vignette$lib.loc
#> NULL
#> 
#> $vignette$all
#> [1] TRUE
#> 
#> 
#> $getS3method
#> $getS3method$f
#> 
#> 
#> $getS3method$class
#> 
#> 
#> $getS3method$optional
#> [1] FALSE
#> 
#> $getS3method$envir
#> parent.frame()
#> 
#> 
#> $askYesNo
#> $askYesNo$msg
#> 
#> 
#> $askYesNo$default
#> [1] TRUE
#> 
#> $askYesNo$prompts
#> getOption("askYesNo", gettext(c("Yes", "No", "Cancel")))
#> 
#> $askYesNo$...
#> 
#> 
#> 
#> $combn
#> $combn$x
#> 
#> 
#> $combn$m
#> 
#> 
#> $combn$FUN
#> NULL
#> 
#> $combn$simplify
#> [1] TRUE
#> 
#> $combn$...
#> 
#> 
#> 
#> $loadhistory
#> $loadhistory$file
#> [1] ".Rhistory"
#> 
#> 
#> $update.packages
#> $update.packages$lib.loc
#> NULL
#> 
#> $update.packages$repos
#> getOption("repos")
#> 
#> $update.packages$contriburl
#> contrib.url(repos, type)
#> 
#> $update.packages$method
#> 
#> 
#> $update.packages$instlib
#> NULL
#> 
#> $update.packages$ask
#> [1] TRUE
#> 
#> $update.packages$available
#> NULL
#> 
#> $update.packages$oldPkgs
#> NULL
#> 
#> $update.packages$...
#> 
#> 
#> $update.packages$checkBuilt
#> [1] FALSE
#> 
#> $update.packages$type
#> getOption("pkgType")
#> 
#> 
#> $glob2rx
#> $glob2rx$pattern
#> 
#> 
#> $glob2rx$trim.head
#> [1] FALSE
#> 
#> $glob2rx$trim.tail
#> [1] TRUE
#> 
#> 
#> $old.packages
#> $old.packages$lib.loc
#> NULL
#> 
#> $old.packages$repos
#> getOption("repos")
#> 
#> $old.packages$contriburl
#> contrib.url(repos, type)
#> 
#> $old.packages$instPkgs
#> installed.packages(lib.loc = lib.loc, ...)
#> 
#> $old.packages$method
#> 
#> 
#> $old.packages$available
#> NULL
#> 
#> $old.packages$checkBuilt
#> [1] FALSE
#> 
#> $old.packages$...
#> 
#> 
#> $old.packages$type
#> getOption("pkgType")
#> 
#> 
#> $gethash
#> $gethash$h
#> 
#> 
#> $gethash$key
#> 
#> 
#> $gethash$nomatch
#> NULL
#> 
#> 
#> $remove.packages
#> $remove.packages$pkgs
#> 
#> 
#> $remove.packages$lib
#> 
#> 
#> 
#> $RtangleFinish
#> $RtangleFinish$object
#> 
#> 
#> $RtangleFinish$error
#> [1] FALSE
#> 
#> 
#> $getParseData
#> $getParseData$x
#> 
#> 
#> $getParseData$includeText
#> [1] NA
#> 
#> 
#> $xedit
#> $xedit$name
#> NULL
#> 
#> $xedit$file
#> [1] ""
#> 
#> 
#> $RtangleRuncode
#> $RtangleRuncode$object
#> 
#> 
#> $RtangleRuncode$chunk
#> 
#> 
#> $RtangleRuncode$options
#> 
#> 
#> 
#> $packageName
#> $packageName$env
#> parent.frame()
#> 
#> 
#> $bibentry
#> $bibentry$bibtype
#> 
#> 
#> $bibentry$textVersion
#> NULL
#> 
#> $bibentry$header
#> NULL
#> 
#> $bibentry$footer
#> NULL
#> 
#> $bibentry$key
#> NULL
#> 
#> $bibentry$...
#> 
#> 
#> $bibentry$other
#> list()
#> 
#> $bibentry$mheader
#> NULL
#> 
#> $bibentry$mfooter
#> NULL
#> 
#> 
#> $RtangleSetup
#> $RtangleSetup$file
#> 
#> 
#> $RtangleSetup$syntax
#> 
#> 
#> $RtangleSetup$output
#> NULL
#> 
#> $RtangleSetup$annotate
#> [1] TRUE
#> 
#> $RtangleSetup$split
#> [1] FALSE
#> 
#> $RtangleSetup$quiet
#> [1] FALSE
#> 
#> $RtangleSetup$drop.evalFALSE
#> [1] FALSE
#> 
#> $RtangleSetup$...
#> 
#> 
#> 
#> $findLineNum
#> $findLineNum$srcfile
#> 
#> 
#> $findLineNum$line
#> 
#> 
#> $findLineNum$nameonly
#> [1] TRUE
#> 
#> $findLineNum$envir
#> parent.frame()
#> 
#> $findLineNum$lastenv
#> 
#> 
#> 
#> $RtangleWritedoc
#> $RtangleWritedoc$object
#> 
#> 
#> $RtangleWritedoc$chunk
#> 
#> 
#> 
#> $txtProgressBar
#> $txtProgressBar$min
#> [1] 0
#> 
#> $txtProgressBar$max
#> [1] 1
#> 
#> $txtProgressBar$initial
#> [1] 0
#> 
#> $txtProgressBar$char
#> [1] "="
#> 
#> $txtProgressBar$width
#> [1] NA
#> 
#> $txtProgressBar$title
#> 
#> 
#> $txtProgressBar$label
#> 
#> 
#> $txtProgressBar$style
#> [1] 1
#> 
#> $txtProgressBar$file
#> [1] ""
#> 
#> 
#> $promptData
#> $promptData$object
#> 
#> 
#> $promptData$filename
#> NULL
#> 
#> $promptData$name
#> NULL
#> 
#> 
#> $file_test
#> $file_test$op
#> 
#> 
#> $file_test$x
#> 
#> 
#> $file_test$y
#> 
#> 
#> 
#> $getSrcFilename
#> $getSrcFilename$x
#> 
#> 
#> $getSrcFilename$full.names
#> [1] FALSE
#> 
#> $getSrcFilename$unique
#> [1] TRUE
#> 
#> 
#> $help
#> $help$topic
#> 
#> 
#> $help$package
#> NULL
#> 
#> $help$lib.loc
#> NULL
#> 
#> $help$verbose
#> getOption("verbose")
#> 
#> $help$try.all.packages
#> getOption("help.try.all.packages")
#> 
#> $help$help_type
#> getOption("help_type")
#> 
#> 
#> $toLatex
#> $toLatex$object
#> 
#> 
#> $toLatex$...
#> 
#> 
#> 
#> $installed.packages
#> $installed.packages$lib.loc
#> NULL
#> 
#> $installed.packages$priority
#> NULL
#> 
#> $installed.packages$noCache
#> [1] FALSE
#> 
#> $installed.packages$fields
#> NULL
#> 
#> $installed.packages$subarch
#> .Platform$r_arch
#> 
#> $installed.packages$...
#> 
#> 
#> 
#> $citFooter
#> $citFooter$...
#> 
#> 
#> 
#> $unstack
#> $unstack$x
#> 
#> 
#> $unstack$...
#> 
#> 
#> 
#> $mirror2html
#> $mirror2html$mirrors
#> NULL
#> 
#> $mirror2html$file
#> [1] "mirrors.html"
#> 
#> $mirror2html$head
#> [1] "mirrors-head.html"
#> 
#> $mirror2html$foot
#> [1] "mirrors-foot.html"
#> 
#> 
#> $debugcall
#> $debugcall$call
#> 
#> 
#> $debugcall$once
#> [1] FALSE
#> 
#> 
#> $data.entry
#> $data.entry$...
#> 
#> 
#> $data.entry$Modes
#> NULL
#> 
#> $data.entry$Names
#> NULL
#> 
#> 
#> $str
#> $str$object
#> 
#> 
#> $str$...
#> 
#> 
#> 
#> $read.fortran
#> $read.fortran$file
#> 
#> 
#> $read.fortran$format
#> 
#> 
#> $read.fortran$...
#> 
#> 
#> $read.fortran$as.is
#> [1] TRUE
#> 
#> $read.fortran$colClasses
#> [1] NA
#> 
#> 
#> $dataentry
#> $dataentry$data
#> 
#> 
#> $dataentry$modes
#> 
#> 
#> 
#> $citation
#> $citation$package
#> [1] "base"
#> 
#> $citation$lib.loc
#> NULL
#> 
#> $citation$auto
#> NULL
#> 
#> 
#> $emacs
#> $emacs$name
#> NULL
#> 
#> $emacs$file
#> [1] ""
#> 
#> 
#> $head
#> $head$x
#> 
#> 
#> $head$...
#> 
#> 
#> 
#> $promptImport
#> $promptImport$object
#> 
#> 
#> $promptImport$filename
#> NULL
#> 
#> $promptImport$name
#> NULL
#> 
#> $promptImport$importedFrom
#> NULL
#> 
#> $promptImport$importPage
#> name
#> 
#> $promptImport$...
#> 
#> 
#> 
#> $removeSource
#> $removeSource$fn
#> 
#> 
#> 
#> $make.packages.html
#> $make.packages.html$lib.loc
#> .libPaths()
#> 
#> $make.packages.html$temp
#> [1] FALSE
#> 
#> $make.packages.html$verbose
#> [1] TRUE
#> 
#> $make.packages.html$docdir
#> R.home("doc")
#> 
#> 
#> $RweaveChunkPrefix
#> $RweaveChunkPrefix$options
#> 
#> 
#> 
#> $page
#> $page$x
#> 
#> 
#> $page$method
#> c("dput", "print")
#> 
#> $page$...
#> 
#> 
#> 
#> $aspell_write_personal_dictionary_file
#> $aspell_write_personal_dictionary_file$x
#> 
#> 
#> $aspell_write_personal_dictionary_file$out
#> 
#> 
#> $aspell_write_personal_dictionary_file$language
#> [1] "en"
#> 
#> $aspell_write_personal_dictionary_file$program
#> NULL
#> 
#> 
#> $formatOL
#> $formatOL$x
#> 
#> 
#> $formatOL$type
#> [1] "arabic"
#> 
#> $formatOL$offset
#> [1] 0
#> 
#> $formatOL$start
#> [1] 1
#> 
#> $formatOL$width
#> 0.9 * getOption("width")
#> 
#> 
#> $RweaveLatexWritedoc
#> $RweaveLatexWritedoc$object
#> 
#> 
#> $RweaveLatexWritedoc$chunk
#> 
#> 
#> 
#> $memory.limit
#> $memory.limit$size
#> [1] NA
#> 
#> 
#> $find
#> $find$what
#> 
#> 
#> $find$mode
#> [1] "any"
#> 
#> $find$numeric
#> [1] FALSE
#> 
#> $find$simple.words
#> [1] TRUE
#> 
#> 
#> $pico
#> $pico$name
#> NULL
#> 
#> $pico$file
#> [1] ""
#> 
#> 
#> $fixInNamespace
#> $fixInNamespace$x
#> 
#> 
#> $fixInNamespace$ns
#> 
#> 
#> $fixInNamespace$pos
#> -1
#> 
#> $fixInNamespace$envir
#> as.environment(pos)
#> 
#> $fixInNamespace$...
#> 
#> 
#> 
#> $adist
#> $adist$x
#> 
#> 
#> $adist$y
#> NULL
#> 
#> $adist$costs
#> NULL
#> 
#> $adist$counts
#> [1] FALSE
#> 
#> $adist$fixed
#> [1] TRUE
#> 
#> $adist$partial
#> !fixed
#> 
#> $adist$ignore.case
#> [1] FALSE
#> 
#> $adist$useBytes
#> [1] FALSE
#> 
#> 
#> $example
#> $example$topic
#> 
#> 
#> $example$package
#> NULL
#> 
#> $example$lib.loc
#> NULL
#> 
#> $example$character.only
#> [1] FALSE
#> 
#> $example$give.lines
#> [1] FALSE
#> 
#> $example$local
#> [1] FALSE
#> 
#> $example$type
#> c("console", "html")
#> 
#> $example$echo
#> [1] TRUE
#> 
#> $example$verbose
#> getOption("verbose")
#> 
#> $example$setRNG
#> [1] FALSE
#> 
#> $example$ask
#> getOption("example.ask")
#> 
#> $example$prompt.prefix
#> abbreviate(topic, 6)
#> 
#> $example$run.dontrun
#> [1] FALSE
#> 
#> $example$run.donttest
#> interactive()
#> 
#> 
#> $findCRANmirror
#> $findCRANmirror$type
#> c("src", "web")
#> 
#> 
#> $data
#> $data$...
#> 
#> 
#> $data$list
#> character()
#> 
#> $data$package
#> NULL
#> 
#> $data$lib.loc
#> NULL
#> 
#> $data$verbose
#> getOption("verbose")
#> 
#> $data$envir
#> .GlobalEnv
#> 
#> $data$overwrite
#> [1] TRUE
#> 
#> 
#> $write.csv2
#> $write.csv2$...
#> 
#> 
#> 
#> $memory.size
#> $memory.size$max
#> [1] FALSE
#> 
#> 
#> $menu
#> $menu$choices
#> 
#> 
#> $menu$graphics
#> [1] FALSE
#> 
#> $menu$title
#> NULL
#> 
#> 
#> $read.csv
#> $read.csv$file
#> 
#> 
#> $read.csv$header
#> [1] TRUE
#> 
#> $read.csv$sep
#> [1] ","
#> 
#> $read.csv$quote
#> [1] "\""
#> 
#> $read.csv$dec
#> [1] "."
#> 
#> $read.csv$fill
#> [1] TRUE
#> 
#> $read.csv$comment.char
#> [1] ""
#> 
#> $read.csv$...
#> 
#> 
#> 
#> $url.show
#> $url.show$url
#> 
#> 
#> $url.show$title
#> url
#> 
#> $url.show$file
#> tempfile()
#> 
#> $url.show$delete.file
#> [1] TRUE
#> 
#> $url.show$method
#> 
#> 
#> $url.show$...
#> 
#> 
#> 
#> $history
#> $history$max.show
#> [1] 25
#> 
#> $history$reverse
#> [1] FALSE
#> 
#> $history$pattern
#> 
#> 
#> $history$...
#> 
#> 
#> 
#> $chooseBioCmirror
#> $chooseBioCmirror$graphics
#> getOption("menu.graphics")
#> 
#> $chooseBioCmirror$ind
#> NULL
#> 
#> $chooseBioCmirror$local.only
#> [1] FALSE
#> 
#> 
#> $View
#> $View$x
#> 
#> 
#> $View$title
#> 
#> 
#> 
#> $assignInMyNamespace
#> $assignInMyNamespace$x
#> 
#> 
#> $assignInMyNamespace$value
#> 
#> 
#> 
#> $hsearch_db
#> $hsearch_db$package
#> NULL
#> 
#> $hsearch_db$lib.loc
#> NULL
#> 
#> $hsearch_db$types
#> getOption("help.search.types")
#> 
#> $hsearch_db$verbose
#> getOption("verbose")
#> 
#> $hsearch_db$rebuild
#> [1] FALSE
#> 
#> $hsearch_db$use_UTF8
#> [1] FALSE
#> 
#> 
#> $browseURL
#> $browseURL$url
#> 
#> 
#> $browseURL$browser
#> getOption("browser")
#> 
#> $browseURL$encodeIfNeeded
#> [1] FALSE
#> 
#> 
#> $untar
#> $untar$tarfile
#> 
#> 
#> $untar$files
#> NULL
#> 
#> $untar$list
#> [1] FALSE
#> 
#> $untar$exdir
#> [1] "."
#> 
#> $untar$compressed
#> [1] NA
#> 
#> $untar$extras
#> NULL
#> 
#> $untar$verbose
#> [1] FALSE
#> 
#> $untar$restore_times
#> [1] TRUE
#> 
#> $untar$support_old_tars
#> Sys.getenv("R_SUPPORT_OLD_TARS", FALSE)
#> 
#> $untar$tar
#> Sys.getenv("TAR")
#> 
#> 
#> $aspell
#> $aspell$files
#> 
#> 
#> $aspell$filter
#> 
#> 
#> $aspell$control
#> list()
#> 
#> $aspell$encoding
#> [1] "unknown"
#> 
#> $aspell$program
#> NULL
#> 
#> $aspell$dictionaries
#> character()
#> 
#> 
#> $zip
#> $zip$zipfile
#> 
#> 
#> $zip$files
#> 
#> 
#> $zip$flags
#> [1] "-r9X"
#> 
#> $zip$extras
#> [1] ""
#> 
#> $zip$zip
#> Sys.getenv("R_ZIPCMD", "zip")
#> 
#> 
#> $make.socket
#> $make.socket$host
#> [1] "localhost"
#> 
#> $make.socket$port
#> 
#> 
#> $make.socket$fail
#> [1] TRUE
#> 
#> $make.socket$server
#> [1] FALSE
#> 
#> 
#> $getSrcDirectory
#> $getSrcDirectory$x
#> 
#> 
#> $getSrcDirectory$unique
#> [1] TRUE
#> 
#> 
#> $suppressForeignCheck
#> $suppressForeignCheck$names
#> 
#> 
#> $suppressForeignCheck$package
#> 
#> 
#> $suppressForeignCheck$add
#> [1] TRUE
#> 
#> 
#> $undebugcall
#> $undebugcall$call
#> 
#> 
#> 
#> $Stangle
#> $Stangle$file
#> 
#> 
#> $Stangle$driver
#> Rtangle()
#> 
#> $Stangle$syntax
#> getOption("SweaveSyntax")
#> 
#> $Stangle$encoding
#> [1] ""
#> 
#> $Stangle$...
#> 
#> 
#> 
#> $aspell_package_C_files
#> $aspell_package_C_files$dir
#> 
#> 
#> $aspell_package_C_files$ignore
#> character()
#> 
#> $aspell_package_C_files$control
#> list()
#> 
#> $aspell_package_C_files$program
#> NULL
#> 
#> $aspell_package_C_files$dictionaries
#> character()
#> 
#> 
#> $download.file
#> $download.file$url
#> 
#> 
#> $download.file$destfile
#> 
#> 
#> $download.file$method
#> 
#> 
#> $download.file$quiet
#> [1] FALSE
#> 
#> $download.file$mode
#> [1] "w"
#> 
#> $download.file$cacheOK
#> [1] TRUE
#> 
#> $download.file$extra
#> getOption("download.file.extra")
#> 
#> $download.file$headers
#> NULL
#> 
#> $download.file$...
#> 
#> 
#> 
#> $apropos
#> $apropos$what
#> 
#> 
#> $apropos$where
#> [1] FALSE
#> 
#> $apropos$ignore.case
#> [1] TRUE
#> 
#> $apropos$mode
#> [1] "any"
#> 
#> 
#> $as.roman
#> $as.roman$x
#> 
#> 
#> 
#> $xemacs
#> $xemacs$name
#> NULL
#> 
#> $xemacs$file
#> [1] ""
#> 
#> 
#> $strcapture
#> $strcapture$pattern
#> 
#> 
#> $strcapture$x
#> 
#> 
#> $strcapture$proto
#> 
#> 
#> $strcapture$perl
#> [1] FALSE
#> 
#> $strcapture$useBytes
#> [1] FALSE
#> 
#> 
#> $methods
#> $methods$generic.function
#> 
#> 
#> $methods$class
#> 
#> 
#> 
#> $typhash
#> $typhash$h
#> 
#> 
#> 
#> $as.person
#> $as.person$x
#> 
#> 
#>